feat(home): organize HomeScreen and add route
- Move `HomeScreen` to the `home` feature package.
- Set up navigation route for the Home screen.
- Add `build_runner` script to `melos` and update `Makefile`
