+++ app/lib/src/splash/splash_route.dart
@@ -10,9 +10,7 @@ class SplashRoute extends GoRoute {
           path: '/',
           builder: (context, state) {
             return SplashScreen(
-              onAnimationDone: () {
-                print('onAnimationDone');
-              },
+              onAnimationDone: () => context.goNamed(RouteNames.home),
             );
           },
         );
