+++ app/pubspec.yaml
@@ -12,17 +12,15 @@ dependencies:
   core: ^0.0.1
   flutter:
     sdk: flutter
+  home: ^0.0.1
   http: ^1.2.2
   i18n: ^0.0.1
-  json_annotation: ^4.9.0
   lottie: ^3.1.3
 
 dev_dependencies:
-  build_runner: ^2.4.10
   flutter_lints: ^4.0.0
   flutter_test:
     sdk: flutter
-  json_serializable: ^6.8.0
 
 flutter:
   generate: true
