+++ common/core/lib/src/navigation/route_names.dart
@@ -3,4 +3,7 @@ abstract final class RouteNames {
 
   // Splash
   static const String splash = 'Splash';
+
+  // Home
+  static const String home = 'Home Screen';
 }
