+++ common/i18n/lib/l10n/en-US.json
@@ -1,5 +1,5 @@
 {
-    "main": {
+    "home": {
         "title": "Restaurant Tour"
     }
 }
