+++ common/i18n/lib/l10n/pt-BR.json
@@ -1,5 +1,5 @@
 {
-    "main": {
+    "home": {
         "title": "Tour de Restaurantes"
     }
 }
