+++ features/home/lib/home.dart
@@ -1,1 +1,3 @@
 library;
+
+export 'src/home_route.dart';
