+++ features/home/lib/src/home_route.dart
@@ -1,0 +1,15 @@
+import 'package:core/core.dart';
+import 'package:core/go_router.dart';
+
+import 'widgets/home_screen.dart';
+
+class HomeRoute extends GoRoute {
+  HomeRoute()
+      : super(
+          name: RouteNames.home,
+          path: '/home',
+          builder: (context, state) {
+            return const HomeScreen();
+          },
+        );
+}
