+++ features/home/pubspec.yaml
@@ -8,13 +8,18 @@ environment:
   flutter: ">=3.0.0"
 
 dependencies:
+  core: ^0.0.1
   flutter:
     sdk: flutter
+  i18n: ^0.0.1
+  json_annotation: ^4.9.0
 
 dev_dependencies:
+  build_runner: ^2.4.10
   flutter_lints: ^4.0.0
   flutter_test:
     sdk: flutter
+  json_serializable: ^6.8.0
 
 flutter:
   uses-material-design: true
